details content placed in a separate folder
