+++ src/App/features/Details/MoviePage/MovieInfo/index.js
