+++ src/App/features/Details/MoviePage/MovieTop/styled.js
