+++ src/App/features/Details/MoviePage/styled.js
